fixed entities

+++ Domain/Entities/OrderItem.cs
@@ -4,7 +4,7 @@ public class OrderItem
     public Guid OrderItemId { get; private init; }
     public Guid OrderId { get; private set; }
     public Guid ProductId { get; private set; }
-    public int Quantity { get; private set; }
+    public decimal Quantity { get; private set; }
     public decimal UnitPrice { get; private set; }
     public Order Order { get; private set; }
     public Product Product { get; private set; }
